LVMv2: initial implementation system(600M)
-Logical Volume Management

-The YP-G70 only has a 290m /system partition and this is not enough
to accomodate KitKat + full gapps. So, I'm implementing LVM so that
we can get dynamic volume resizing without having to change the under-
lying partitions.

-Partitions in Logical Group (Aprox 2.1G)
-- /dev/block/mmcblk0p13 (normally /system)
-- /dev/block/mmcblk0p16 (normally /data

-Logical volumes
-- /dev/lvpool/system (600M)
-- /dev/lvpool/data (100% remaining, aprox 1.5G)

-Installation Notes
--On the first flash, if the system logical volume is not present or
is less than 600M there is an assert error. This is so that if anyone
needs to backup information contained in /data, they can. In order to
proceed with the installation, the zip must be flashed a 2nd time.
--During the 2nd flash partitions in the logical group are wiped and
the logical volumes are created and formated then the /etc/fstab and
/etc/recovery.fstab are switched so that, when we hand everything off
to the normal rom installer, the new logical volumes are mounted.

-Reverting to a PRE-LVM ROM
--In the YP-G70 the "recovery", be it CWM or TWRP and in the kernel's
recovery-ramdisk. So in order to get a different recovery you just
flash a different kernel. So to get back to a PRE-LVM ROM, you just:
1. Reboot to recovery like normal.
2. Flash your PRE-LVM ROM like normal.
-- Note: All we really care about at this point is that the kernel
(boot.img) gets flashed. Ignore any other errors!
3. After the flash, Power off.
4. Hold volume up and power on, till you get to recovery.
5. Full wipe, including /system
6. Flash ROM + Gapps like normal

-Future Considerations
--LVMv2 allows us to quickly change the size of the system partition
as Android grows in size without having to manually repartition.
Manually repartitioning might cause unforseen problems with legacy
software.

-Meticulus

Conflicts:
	BoardConfigCommon.mk
	etc/twrp.fstab