camera HAL: Reworked to give rear cam photo snap limited functionality.
-This does not fix the rear camera and a proper solution should still be
persued.

-A proper solution to the rear camera photosnap problem does not seem
forthcomming so in order to give the rear cam photo snap some functionality;
It's been reduced to a vga camera.It picks images off of the preview stream,
just as the front cam does. This is similar to the way that apps such as
FastBurst operate.

-The benefits of this implementation are:
 1. The rear cam can snap photo's in *any app.

-The trade off of this implementation are:
 1.  Photo resolution is limited to 800x480 or 640x480

-The "real" problem seems to be that when we ask the rear cam to
"switch" to the "high resolution jpeg mode" we get an error from
a really low level in kernel code...

* = theoretically