LMK-NKP: Enable and don't kill wpa_supplicant.
-A look into the "wifi" issue, in which the wifi will not
toggle on (in rare instances), shows that wpa_supplicant
is not running. There is a mechanism to restart wpa_supplicant
when the error occurs and the process does work but it seems
that the front ie(system ui and Settings) are never notified
that the "wpa_supplicant" restart happens so it appears "hung"
until a restart.

-I suspect that wpa_supplicant is killed due to long periods of
inactivity and memory demands on our low memory device. This
should cause the LMK not to kill wpa_supplicant unless absolutely
neccessary.